% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{dat}
\alias{dat}
\title{Simulated data for demonstrating \code{"fglsnet"}.}
\format{
An object of class \code{list} of length 3.
}
\usage{
data(dat)
}
\description{
Simulated data for demonstrating \code{"fglsnet"}.
}
\details{
\code{Y} is the outcome. \code{X} contains the regressors including the intercept.. \code{M} is the dependence network.
}
\keyword{datasets}
